+++ packages/backend-storage/src/construct.test.ts
@@ -97,6 +97,7 @@ void describe('AmplifyStorage', () => {
       const storageStrategy: BackendOutputStorageStrategy<BackendOutputEntry> =
         {
           addBackendOutputEntry: storeOutputMock,
+          appendToBackendOutputList: storeOutputMock,
         };
 
       const storageConstruct = new AmplifyStorage(stack, 'test', {